pyproject.toml
[tool.poetry]
name = "igreja"
version = "0.1.0"
description = ""
authors = ["Danilo da Conceição Nascimento <daconnas.dcn@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.12"
Django = "^4.0.3"
django-model-utils = "^4.2.0"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-ckeditor = "^6.2.0"
dj-database-url = "^0.5.0"
python-decouple = "^3.6"
psycopg2 = "^2.9.3"
whitenoise = "^6.0.0"
gunicorn = "^20.1.0"

[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
isort = "^5.10.1"
black = "^22.1.0"
mypy = "^0.942"
tox = "^3.24.5"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
